requirements.txt
grpcio>=1.48.0
grpcio-tools>=1.48.0
protobuf>=4.0.0